Move AttachedPresentation to gaphor.diagram.presentation
